Updated code to adhere to ADR018 (#270)
## Description

This contains a breaking change, since the version that is specified for NiFi now has to include a stackable image version.

Previously the version could be specified as "version: 3.2.1" and the operator simply added the stackable version to this. But going forward we have decided that the image version has to be explictly specified, so it will now have to be "version: 3.2.1-stackable-0.1.0" or similar.

This is in accordance with ADR18 (https://docs.stackable.tech/home/contributor/adr/ADR018-product_image_versioning.html).

This PR also moves the integration tests back to this repository and parameterizes them to test with our currently supported ZooKeeper versions.