Initial attempt at making resource requests, limits and pvcs configurable (#265)
## Description
Fixes #252

This doesn't work with the currently released version of the operator-rs, and is only intended for code study, as this might serve as a template for making similar changes in other operators.

Most notably it depends on https://github.com/stackabletech/operator-rs/pull/368 and a PR to make small changes to the Resource structs to make them work with the Merging code, which can in turn only be created once the above mentioned PR has been merged.


*Please add a description here. This will become the commit message of the merge request later.*



Co-authored-by: Razvan-Daniel Mihai <84674+razvan@users.noreply.github.com>