Getting started (#330)
# Description

There is a long test script in the examples/code. Not quite happy about that but it works fine. Maybe a little too complicated for a getting started guide though. Any thoughts on that? Maybe just leave it for testing the script, but not expose it in the docs?

fixes https://github.com/stackabletech/nifi-operator/issues/325



Co-authored-by: Malte Sander <malte.sander.it@gmail.com>