Adjust NiFi repository sizes depending on configured PVC sizes (#371)
# Description

Fixes #354

Currently this reserves a constant 90% of the PVC size. In practice this headroom "should" probably be scaled based on the complexity on the NiFi graph, but that's not really a thing we can predict easily.

This has been running over the weekend with a test cluster and it has cleaned up and stayed working over that period, with all volumes sitting around their respective fill points.